Fix for M1 Apple hardware
This is a WIP, and builds upon PR #469.

It updates protoc and grpc to latest so we get aarch_64 library support, but then
this changed service URIs to `svc:///service-name` from simply `service-name` before.

As I am a coward, this change checks that it's in svc:/// format, and falls back to
the old behaviour if it isn't.  This may not be necessary.

The entirity of the change (apart from version bumping, and test tweaks to give me a
greater sense of trust) is in the file:

io.micronaut.grpc.discovery.GrpcNameResolverProvider#newNameResolver