Fix round robin load balancing with service discovery
GrpcNameResolverProvider maps individual ServiceInstances provided by
the discovery client 1-to-1 to instances of EquivalentAddressGroup
instead of creating a single EquivalentAddressGroup instance with a
list of addresses, as the GRPC load balancing algorithms are designed to
balance across these group instances.

A test is added to verify that the load balancing works correctly.

This resolves #818.